+++ internal/volume/pdf.go
@@ -54,6 +54,11 @@ func (s *pdfSource) PageCount() int {
 	return s.pageCount
 }
 
+// Skipped is always zero because every page in a pdf file is an image page
+func (s *pdfSource) Skipped() SkippedPages {
+	return SkippedPages{}
+}
+
 func (s *pdfSource) Metadata() SourceMetadata {
 	return SourceMetadata{}
 }